Create refresh.php
the first if loop grabs the session file timestamp and sends it over to the javascript file handling the popup. the second section is when the user refreshes their session with the popup. essentially to refresh a php session, just make a server request with that users cookie. the array of details is used by another page that shows active users.